backend/requirements.txt
# Django 核心（固定框架版本，稳定）
asgiref==3.11.1
certifi==2026.2.25
charset-normalizer==3.4.6
Django==5.2.12
django-cors-headers==4.9.0
django-filter==24.3
django-redis==6.0.0
django-simpleui==2026.1.13
django-stubs==5.1.0
django-stubs-ext==5.1.0
djangorestframework==3.17.1
djangorestframework-stubs==3.15.0
djangorestframework-simplejwt==5.5.1
drf-spectacular==0.27.2
idna==3.11
Markdown==3.10.2
mysqlclient==2.2.8
pillow==11.0.0
PyJWT==2.12.1
PyMySQL==1.1.2
redis==5.2.0
requests==2.33.0
sqlparse==0.5.5
types-PyYAML==6.0.12.20240917
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.6.3

# LangChain 核心包 - 严格使用 1.0.3 版本（与源项目保持一致）
langchain==1.0.3
langchain-core==1.0.3
langchain-openai==1.0.2
langchain-community==0.4.1

# LangGraph - 用于构建复杂工作流
langgraph==1.0.2
langgraph-checkpoint-sqlite==3.0.0

# LangChain 集成包
langchain-tavily==0.2.13
langchain-text-splitters==1.0.0

# 向量数据库
faiss-cpu==1.12.0

# 文档加载器
pypdf==5.1.0
unstructured==0.18.15
beautifulsoup4==4.12.3
lxml==5.3.0

# 文件处理
python-multipart==0.0.20
aiofiles==24.1.0

# 数据验证 - Pydantic 2
pydantic==2.12.4
pydantic-settings==2.11.0

# HTTP 客户端
httpx[socks]==0.28.1

# 环境变量
python-dotenv==1.0.1

# 工具相关
tavily-python==0.5.0

# 日志和监控
loguru==0.7.3

# CLI 工具
click==8.1.8
rich==13.9.4